Eu te amo Mummi!! Feliz 62 anos!!
at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil
eco folder
｛
｝
                        -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/draw2d_gdiplus/draw2d_gdiplus_graphics.cpp
@@ -1093,7 +1093,7 @@ gdi_fallback:
    bool graphics::TextOut(int32_t x, int32_t y, const string & str)
    { 
       
-      return TextOut((double) x, (double) y, str, str.get_length()); 
+      return TextOut((double) x, (double) y, str, (int32_t) str.get_length()); 
    
    }
